move the util.py to psci
